Fixed this beauty

+++ main/src/gps.cpp
@@ -6,18 +6,18 @@ namespace gps {
 
 // Private
 namespace {
-    constexpr std::uint8_t BAUD_RATE = 9600;
-    constexpr std::uint8_t RX_PIN = 34;
-    constexpr std::uint8_t TX_PIN = 12;
-    constexpr std::uint8_t SERIAL_NUM = 1;
+constexpr std::uint16_t BAUD_RATE = 9600;
+constexpr std::uint8_t RX_PIN = 34;
+constexpr std::uint8_t TX_PIN = 12;
+constexpr std::uint8_t SERIAL_NUM = 1;
 
-    TinyGPSPlus tiny_gps;
+TinyGPSPlus tiny_gps;
 
-    void print_debug(const Data& gps_data) {
-        Serial.printf("Lat: %.6f Long: %.6f Time: %02d:%02d:%02d\n",
-            gps_data.lat, gps_data.lng, gps_data.time.hours,
-            gps_data.time.minutes, gps_data.time.seconds);
-    }
+void print_debug(const Data& gps_data) {
+    Serial.printf("Lat: %.6f Long: %.6f Time: %02d:%02d:%02d\n",
+        gps_data.lat, gps_data.lng, gps_data.time.hours,
+        gps_data.time.minutes, gps_data.time.seconds);
+}
 } // namespace
 
 HardwareSerial GPSSerial(SERIAL_NUM);
@@ -45,7 +45,6 @@ void gps_task([[maybe_unused]] void *pvParameters) {
 #endif
             }
         }
-
         vTaskDelay(pdMS_TO_TICKS(100));
     }
 }
